fix integration test
