+++ packages/vaultage-client/integration-test/tsconfig.json
@@ -1,0 +1,20 @@
+{
+    "extends": "../../../tsconfig.json",
+    "compilerOptions": {
+      "outDir": "../dist",
+      "baseUrl": "../..",
+      "declaration": true,
+      "sourceMap": true,
+      "lib": [ "es2015", "dom" ],
+      "module": "commonjs",
+      "moduleResolution": "node",
+      "target": "es5"
+    },
+    "include": [
+      "integration-test.ts",
+    ],
+    "exclude": [
+      "node_modules",
+      "dist"
+    ]
+  }
